+++ todo_app/data/trello_items.py
@@ -1,0 +1,25 @@
+import os
+import requests
+get_items_url = "https://api.trello.com/1/boards/6212689e46f57218f07af552/lists/open"
+
+def get_items():
+    items = []
+    querystring = { "key": os.getenv("TRELLO_BOARD_KEY"),
+                    "token":os.getenv("TRELLO_BOARD_TOKEN"),
+                    "cards":"open"}
+    response = requests.request("GET", get_items_url, params=querystring)
+    response_json = response.json()     
+    for trello_list in response_json:
+        for card in trello_list['cards']:
+            card['status'] = trello_list['name']
+            card['title'] = card ['name']
+            items.append(card)
+    return items
+
+def add_item(title):
+    # ToDo
+    return None
+
+def save_item(item):
+    # ToDo
+    return None
